Implement Badge Updater
This introduces:
* New dependency: requests
* New root package: badgeupdater
* Modification in the Badge Admin site Change form
* TestBadgeClients (which assigns a badge to all exisiting Players)

The communication uses conveniently defined models for data interchange
(with accompanying Serializer classes) as well as HMAC for verifying
message integrity and authentication.

Third-parties can use /badgeupdater/client/ (as well as common files
inside /badgeupdater/) to create their own clients, or just implement
the protocol (which is as easy as simple JSON parsing/serializing and
HMAC verification). The database can be accessed using the REST API.

First-parties can use convenient class called LocalBadgeClient and
access the database directly, without the API as a middleware.